Add files via upload
utility.py is python code that has the Basel equations code and code. Some tweaking still required.
bank_run.py is python code that has bank run simulations. It is a monte carlo simulation that goes through different amounts withdrawn during it and whether the bank will be able to cover it, either with cash on hand or selling of assets.